// src/app/model/loanAccountsMapping.ts
export class LoanAccountsMapping {
  id: number | undefined;
  database1: string | undefined;
  company1: string | undefined;
  account1: string | undefined;
  currency1: string | undefined;
  account_id1: string | undefined;
  database2: string | undefined;
  company2: string | undefined;
  account2: string | undefined;
  currency2: string | undefined;
  account_id2: string | undefined;
  // createdAt string | undefined;
  // updatedAt string | undefined
}